move tests to folders
