+++ Src/MoneyFox.Core.Tests/MoneyFox.Core.Tests.csproj
@@ -38,4 +38,8 @@
         </AssemblyAttribute>
     </ItemGroup>
 
+
+
+
+
 </Project>
