+++ Src/MoneyFox.Core.Tests/Queries/Statistics/GetCategoryProgression/GetCategoryProgressionHandlerTests.cs
@@ -1,4 +1,4 @@
-﻿namespace MoneyFox.Core.Tests.Queries.Statistics;
+﻿namespace MoneyFox.Core.Tests.Queries.Statistics.GetCategoryProgression;
 
 using Core.Common.Settings;
 using Core.Queries.Statistics;
